Move base into pages
