+++ pt_os_web_portal/miniscreen_onboarding/menu_pages/pages/base/_base.py
@@ -1,4 +1,4 @@
-from ..attr.speeds import DEFAULT_INTERVAL
+from ...attr.speeds import DEFAULT_INTERVAL
 
 
 class MenuPageBase:
