+++ pt_os_web_portal/miniscreen_onboarding/menu_pages/pages/carry_on.py
@@ -1,7 +1,7 @@
 from ...menus import Menus
 from ..attr.margins import FIRST_LINE_Y, SECOND_LINE_Y, THIRD_LINE_Y
-from ..base._title_base import TitleMenuPage
 from ..render.helpers import draw_text
+from .base._title_base import TitleMenuPage
 
 
 class CarryOnMenuPage(TitleMenuPage):
